pyproject.toml
[project]
name = "react-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.7",
    "langchain-tavily>=0.2.11",
    "python-dotenv>=1.1.1",
]
